ingestion/pyproject.toml
[project]
name = "lenny-ingestion"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.2",
    "youtube-transcript-api>=0.6.1",
    "apify-client>=1.6.0",
    "langchain-text-splitters>=0.0.1",
    "tqdm>=4.66.0",
    "pandas>=2.0.0",
    "feedparser>=6.0.12",
    "google-api-python-client>=2.187.0",
]
